+++ ConsentViewController/Classes/SourcePointClient/SPNativeUI.swift
@@ -163,16 +163,16 @@ class SPNativeButton: SPNativeUI {
 
 class SPNativeImage: SPNativeUI {
     class Settings: SPNativeUISettings {
-        let src: String
-
+        let src: URL?
+        
         required init(from decoder: Decoder) throws {
             let container = try decoder.container(keyedBy: Keys.self)
-            src = try container.decode(String.self, forKey: .src)
+            src = try? container.decodeIfPresent(URL.self, forKey: .url) ?? (try? container.decodeIfPresent(URL.self, forKey: .src))
             try super.init(from: decoder)
         }
 
         enum Keys: CodingKey {
-            case src
+            case src, url
         }
     }
 
@@ -191,14 +191,15 @@ class SPNativeImage: SPNativeUI {
 
 class SPNativeLongButton: SPNativeUI {
     class Settings: SPNativeUISettings {
-        let onText, offText, customText, text: String
+        let onText, offText, customText: String
+        let text: String?
 
         required init(from decoder: Decoder) throws {
             let container = try decoder.container(keyedBy: Keys.self)
             onText = try container.decode(String.self, forKey: .onText)
             offText = try container.decode(String.self, forKey: .offText)
             customText = try container.decode(String.self, forKey: .customText)
-            text = try container.decode(String.self, forKey: .text)
+            text = try? container.decodeIfPresent(String.self, forKey: .text)
             try super.init(from: decoder)
         }