+++ ConsentViewController/Classes/Views/tvOS/NativePrivacyManager/Common/LongButtonViewCell.swift
@@ -28,6 +28,9 @@ class LongButtonViewCell: UITableViewCell {
 
     func loadUI() {
         label.text = labelText
+        label.setDefaultTextColorForDarkMode()
+        customLabel.setDefaultTextColorForDarkMode()
+        stateLabel.setDefaultTextColorForDarkMode()
         if let customText = customText {
             customLabel.isHidden = false
             customLabel.text = customText