+++ Example/ConsentViewController_ExampleTests/GDPRMessageSpec.swift
@@ -15,18 +15,17 @@ class GDPRMessageSpec: QuickSpec {
     override func spec() {
 
         describe("Test GDPRMessage") {
-
-            var attributeStyle: AttributeStyle?
-            var messageAttribute: MessageAttribute?
-            var messageAction: MessageAction?
-            var gdprMessage: GDPRMessage?
+            var attributeStyle: SPNativeMessage.AttributeStyle?
+            var messageAttribute: SPNativeMessage.Attribute?
+            var messageAction: SPNativeMessage.Action?
+            var gdprMessage: SPNativeMessage?
             let customFileds = ["Custom": "Fileds"]
 
             beforeEach {
-                attributeStyle = AttributeStyle(fontFamily: "System-Font", fontSize: 14, color: "#00FA9A", backgroundColor: "#944488")
-                messageAttribute = MessageAttribute(text: "Test GDPR Message", style: attributeStyle!, customFields: customFileds)
-                messageAction = MessageAction(text: "Test GDPR Message", style: attributeStyle!, customFields: customFileds, choiceId: 12, choiceType: SPActionType.AcceptAll)
-                gdprMessage = GDPRMessage(title: messageAttribute!, body: messageAttribute!, actions: [messageAction!], customFields: customFileds)
+                attributeStyle = SPNativeMessage.AttributeStyle(fontFamily: "System-Font", fontSize: 14, color: "#00FA9A", backgroundColor: "#944488")
+                messageAttribute = SPNativeMessage.Attribute(text: "Test GDPR Message", style: attributeStyle!, customFields: customFileds)
+                messageAction = SPNativeMessage.Action(text: "Test GDPR Message", style: attributeStyle!, customFields: customFileds, choiceType: SPActionType.AcceptAll, url: nil)
+                gdprMessage = SPNativeMessage(title: messageAttribute!, body: messageAttribute!, actions: [messageAction!], customFields: customFileds)
             }
 
             it("Test AttributeStyle method") {
@@ -43,38 +42,6 @@ class GDPRMessageSpec: QuickSpec {
 
             it("Test GDPRMessage method") {
                 expect(gdprMessage?.title.style.color).to(equal("#00FA9A"))
-            }
-
-            context("Test JSON Encoding") {
-                var actual: GDPRMessage?
-                beforeEach {
-                    guard let data = try? JSONEncoder().encodeResult(gdprMessage).get() else {
-                        fail("Failed to encode \(String(describing: GDPRMessage.self))")
-                        return
-                    }
-
-                    actual = try? JSONDecoder().decode(GDPRMessage.self, from: data).get()
-                }
-
-                it("has matching title text") {
-                    expect(actual?.title.text).to(equal(gdprMessage?.title.text))
-                }
-
-                it("has matching body text") {
-                    expect(actual?.body.text).to(equal(gdprMessage?.body.text))
-                }
-
-                it("has matching action choice type") {
-                    expect(actual?.actions.first?.choiceType).to(equal(gdprMessage?.actions.first?.choiceType))
-                }
-
-                it("has matching action choice id") {
-                    expect(actual?.actions.first?.choiceId).to(equal(gdprMessage?.actions.first?.choiceId))
-                }
-
-                it("has matching action custom fields") {
-                    expect(actual?.customFields).to(equal(gdprMessage?.customFields))
-                }
             }
         }
     }