+++ Example/ConsentViewController_ExampleTests/Helpers/GDPRUIiDelegateMock.swift
@@ -9,22 +9,38 @@
 import Foundation
 @testable import ConsentViewController
 
-class GDPRUIDelegateMock: GDPRMessageUIDelegate {
-    weak var consentDelegate: SPDelegate?
-    var loadMessageCalled = false
-    var loadPrivacyManagerCalled = false
-    var loadMessageCalledWith: URL?
-
-    init(_ consentDelegate: SPDelegate) {
-        self.consentDelegate = consentDelegate
-    }
-
-    func loadMessage(fromUrl url: URL) {
-        loadMessageCalled = true
-        loadMessageCalledWith = url
-    }
-
-    func loadPrivacyManager() {
-        loadPrivacyManagerCalled = true
-    }
-}
+//  class GDPRUIDelegateMock: SPMessageUIDelegate {
+//    weak var consentDelegate: SPDelegate?
+//    var loadMessageCalled = false
+//    var loadPrivacyManagerCalled = false
+//    var loadMessageCalledWith: URL?
+//
+//    init(_ consentDelegate: SPDelegate) {
+//        self.consentDelegate = consentDelegate
+//    }
+//
+//    func loadMessage(fromUrl url: URL) {
+//        loadMessageCalled = true
+//        loadMessageCalledWith = url
+//    }
+//
+//    func loadPrivacyManager() {
+//        loadPrivacyManagerCalled = true
+//    }
+//
+//    func loaded(_ controller: UIViewController) {
+//        <#code#>
+//    }
+//
+//    func action(_ action: SPAction, from controller: UIViewController) {
+//        <#code#>
+//    }
+//
+//    func onError(_ error: SPError) {
+//        <#code#>
+//    }
+//
+//    func finished(_ vcFinished: UIViewController) {
+//        <#code#>
+//    }
+//  }