+++ Example/SPGDPRExampleAppUITests/ExampleApp.swift
@@ -12,94 +12,105 @@ import Nimble
 protocol App {
     func launch()
     func terminate()
-    func relaunch(clean: Bool)
+    func relaunch(clean: Bool, resetAtt: Bool)
 }
 
 extension XCUIApplication: App {
-    func relaunch(clean: Bool = false) {
+    func relaunch(clean: Bool = false, resetAtt: Bool = false) {
         UserDefaults.standard.synchronize()
         self.terminate()
+        if #available(iOS 15.0, *), resetAtt {
+            resetAuthorizationStatus(for: .userTracking)
+        }
         clean ?
-            launchArguments.append("-cleanAppsData") :
-            launchArguments.removeAll { $0 == "-cleanAppsData" }
+        launchArguments.append("-cleanAppsData") :
+        launchArguments.removeAll { $0 == "-cleanAppsData" }
         launch()
     }
 }
 
-protocol GDPRUI {
-    var consentUI: XCUIElement { get }
-    var privacyManager: XCUIElement { get }
-    var consentMessage: XCUIElement { get }
+class FirstLayerMessage: XCUIApplication {
+    var messageTitle: XCUIElement { staticTexts["Message Title"].firstMatch }
+    var acceptButton: XCUIElement { buttons["Accept All"].firstMatch }
+    var rejectButton: XCUIElement { buttons["Reject All"].firstMatch }
+    var showOptionsButton: XCUIElement { buttons["Show Options"].firstMatch }
 }
 
 class ExampleApp: XCUIApplication {
-    var privacySettingsButton: XCUIElement {
-        buttons["privacySettingsButton"].firstMatch
+    class GDPRPrivacyManager: XCUIApplication {
+        private var container: XCUIElement {
+            webViews.containing(NSPredicate(format: "label CONTAINS[cd] 'GDPR Privacy Manager'")).firstMatch
+        }
+
+        var messageTitle: XCUIElement {
+            container
+        }
+
+        var purposesTab: XCUIElement {
+            staticTexts.containing(NSPredicate(format: "label CONTAINS[cd] 'PURPOSES'")).firstMatch
+        }
+
+        var acceptAllButton: XCUIElement {
+            container.buttons["Accept All"].firstMatch
+        }
+
+        var rejectAllButton: XCUIElement {
+            container.buttons["Reject All"].firstMatch
+        }
+
+        var saveAndExitButton: XCUIElement {
+            container.buttons["Save & Exit"].firstMatch
+        }
+
+        var cancelButton: XCUIElement {
+            container.buttons["Cancel"].firstMatch
+        }
     }
 
-    var clearConsentButton: XCUIElement {
-        buttons["clearConsentButton"].firstMatch
+    class GDPRMessage: FirstLayerMessage {
+        override var messageTitle: XCUIElement {
+            staticTexts["GDPR Message"].firstMatch
+        }
     }
 
-    var acceptVendorXButton: XCUIElement {
-        buttons["acceptVendorXButton"].firstMatch
+    class CCPAMessage: FirstLayerMessage {
+        override var messageTitle: XCUIElement {
+            staticTexts["CCPA Message"].firstMatch
+        }
     }
 
-    var vendorXConsentStatus: String {
-        staticTexts["vendorXConsentStatus"].label
-    }
-}
+    class ATTPrePrompt: XCUIApplication {
+        private let system = XCUIApplication(bundleIdentifier: "com.apple.springboard")
 
-extension ExampleApp: GDPRUI {
-    var consentUI: XCUIElement {
-        webViews.containing(NSPredicate(format: "(label CONTAINS[cd] 'GDPR Message') OR (label CONTAINS[cd] 'Privacy Center')")).firstMatch
-    }
+        private var container: XCUIElement {
+            webViews.containing(NSPredicate(format: "label CONTAINS[cd] 'ATT pre-prompt'")).firstMatch
+        }
 
-    var privacyManager: XCUIElement {
-        webViews.containing(NSPredicate(format: "label CONTAINS[cd] 'Privacy Center'")).firstMatch
-    }
+        var okButton: XCUIElement {
+            container.buttons["Bring it on"].firstMatch
+        }
 
-    var consentMessage: XCUIElement {
-        webViews.containing(NSPredicate(format: "label CONTAINS[cd] 'GDPR Message'")).firstMatch
-    }
+        private var attAlert: XCUIElement {
+            system.alerts.containing(NSPredicate(format: "label CONTAINS[cd] 'track you'")).firstMatch
+        }
 
-    var termsAndConditionsWebPageTitle: XCUIElement {
-        webViews.containing(NSPredicate(format: "label CONTAINS[cd] 'Address'")).firstMatch
+        var attAlertAllowButton: XCUIElement {
+            attAlert.buttons["Allow"].firstMatch
+        }
     }
 
-    var acceptAllButton: XCUIElement {
-        consentUI.buttons.containing(NSPredicate(format: "label CONTAINS[cd] 'Accept'")).firstMatch
+    let gdprMessage = GDPRMessage()
+    let ccpaMessage = CCPAMessage()
+    let gdprPM = GDPRPrivacyManager()
+    let attPrePrompt = ATTPrePrompt()
+
+    var gdprPrivacyManagerButton: XCUIElement {
+        buttons["GDPR Privacy Manager"].firstMatch
     }
 
-    var rejectAllButton: XCUIElement {
-        consentUI.buttons.containing(NSPredicate(format: "label CONTAINS[cd] 'Reject'")).firstMatch
-    }
-
-    var showOptionsButton: XCUIElement {
-        consentUI.buttons["Show Options"].firstMatch
-    }
-
-    var saveAndExitButton: XCUIElement {
-        consentUI.buttons["Save & Exit"].firstMatch
-    }
-
-    var purposesButton: XCUIElement {
-        staticTexts["PURPOSES"].firstMatch
-    }
-
-    var featuresTab: XCUIElement {
-        staticTexts["Features are a use of the data that you have already agreed to share with us"].firstMatch
-    }
-
-    var termsAndConditionsLink: XCUIElement {
-        consentUI.links["Privacy Policy"].firstMatch
-    }
-
-    var DeviceInformationSwitch: XCUIElement {
-        consentUI.switches["Store and/or access information on a device"].firstMatch
-    }
-
-    var PersonalisedAdsSwitch: XCUIElement {
-        consentUI.switches["Create a personalised ads profile"].firstMatch
+    var shouldRunAttScenario: Bool {
+        /// Unfortunately querying for `ATTrackingManager.trackingAuthorizationStatus` during tests is not reliable.
+        /// So we rely on the app's IDFA status label in order to decide if the ATT scenario should be tested or not.
+        staticTexts["unknown"].exists
     }
 }