+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 6.3.1 (Nov, 17, 2021)
+* Added a new action type, `Custom`. You can be notified when an action is taken by the user via the `onAction` callback, the `Custom` action can be used to trigger any custom flow you have on your app, eg. subscribing, paywall, etc.
+* Added a section on the `SPDelegate` methods to our README.
+* Better document (XCode comments) the `SPUserData` and `SPGDPRConsent` classes.
+* Added a comment to make it more transparent the fact that `SPSDK#userData` accesses `UserDefaults` every time.
+
+# 6.3.0 (Oct, 26, 2021)
+* Good news, the Native Message is back! For more info make sure to check the [Native Message wiki](https://github.com/SourcePointUSA/ios-cmp-app/wiki/Rendering-consent-message-natively) and our [README](https://github.com/SourcePointUSA/ios-cmp-app#readme).
+
+# 6.2.1 (Oct, 20, 2021)
+* Fix an issue preventing the SDK from being built via SPM
+* Add example on how to use vendor grants on README
+
+# 6.2.0 (August, 25, 2021)
+* Officially added support to AppleTV (check our NativePMExample app).
+* The `SPDelegate` protocol has been updated. All methods that expected a `SPMessageViewController` now receive a plain `UIViewController` instead.
+
 # 6.1.7 (August, 25, 2021)
 * Added `acceptedCategories` property to `SPGDPRConsent` class. The `acceptedCategories` is an array of ids of the categories (purposes) consented by the user and that are consented in all vendors involved with it. If a single vendor of a given purpose is disabled by the user, that purpose won't be part of `acceptedCategories`. [#335](https://github.com/SourcePointUSA/ios-cmp-app/pull/335)
 * Fixed an issue preventing using the SDK with Carthage. [#334](https://github.com/SourcePointUSA/ios-cmp-app/pull/334)